JETSON/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.4
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.23.5
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.2.0
pkg_resources==0.0.0
platformdirs==4.3.6
pycuda==2025.1.1
pyparsing==3.1.4
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
pytools==2024.1.14
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch @ file:///home/dawood/Desktop/william/torch-2.1.0-cp38-cp38-linux_aarch64.whl#sha256=5112e2ef5051f1003ae2ffb545ae596377df66979d62a954d774f18009064dbe
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
zipp==3.20.2
